feat: create page A, B, C
