+++ src/pages/A/A.jsx
@@ -1,0 +1,7 @@
+import { Contents } from "./A.styled";
+
+function A() {
+  return <Contents>This is Page A!</Contents>;
+}
+
+export default A;
