+++ src/pages/A/A.styled.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const Contents = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-left: 1px solid black;
+  font-size: 2rem;
+  font-weight: bold;
+`;
+
+export { Contents };
