+++ src/pages/B/B.jsx
@@ -1,0 +1,7 @@
+import { Contents } from "../A/A.styled";
+
+function B() {
+  return <Contents>This is Page B!</Contents>;
+}
+
+export default B;
