+++ src/pages/C/C.jsx
@@ -1,0 +1,7 @@
+import { Contents } from "../A/A.styled";
+
+function C() {
+  return <Contents>This is Page C!</Contents>;
+}
+
+export default C;
